### 2019-09-19 #1748 @whedon
[PRE REVIEW]: widgyts: Custom Jupyter Widgets for Interactive Data Exploration with yt
**Submitting author:** @munkm (<a href="http://orcid.org/0000-0003-0117-5366">Madicken Munk</a>)
**Repository:** <a href="https://github.com/data-exp-lab/widgyts" target ="_blank">https://github.com/data-exp-lab/widgyts</a>
**Version:** v0.3.1
