### 2019-09-19 #1748 @kthyng
Hi @lpantano! Would you be willing to edit this data viz and exploration based submission?
