### 2019-09-23 60a0cc6
add link to hosted documentation in README
