### 2019-09-30 #1748 @lpantano
It was an error, I fixed it. Sorry about that.

On September 30, 2019 at 10:18:16 AM, Madicken Munk (
notifications@github.com) wrote:

@lpantano <https://github.com/lpantano> am I supposed to be a reviewer on
my own paper?
