### 2019-10-23 data-exp-lab/widgyts#45 @harpolea
JOSS paper
I have a couple of suggestions for the JOSS paper submission (https://github.com/openjournals/joss-reviews/issues/1774)

1. One of the points in the review checklist asks for a demonstration of the performance. In the paper, you describe how `widgyts` allows for more efficient exploration of datasets by shifting the image calculation to the client. Do you have any specific examples that demonstrate this?
2. Another of the points asks for a description of the state of the field, including a description of how this compares to other commonly used software packages. Could you add something along these lines to the paper?
