### 2019-10-23 #1774 @harpolea
I'm so sorry for the delay in finishing this!

Overall, this is a really cool package. I've read over the documentation and played around with the examples notebooks, and everything there looks good to me. The python test suite has good coverage (though the coverage reports on `codecov` don't seem to be working at the moment?).

I have a couple of small suggestions for the paper before I tick off the last couple of points in the checklist, which I've listed in this issue: https://github.com/data-exp-lab/widgyts/issues/45
