### 2019-11-11 #1774 @KayleighRutherford
Hello, I did complete the checklist a while back - is there something more
I need to do/how do I close the issue?

On Mon, Nov 11, 2019 at 9:28 AM Lorena Pantano <notifications@github.com>
wrote:

> @KayleighRutherford <https://github.com/KayleighRutherford> do you think
> you can finish the review soon?
>
