### 2019-11-11 #1774 @lpantano
thanks for the reply @KayleighRutherford , I see some of the items not checked under your name checklist, can you double check it is complete, maybe it didn't sync when you did it. Thanks!
