### 2019-12-02 data-exp-lab/widgyts#45 @KayleighRutherford
Hi, just checking in after reviewing this (openjournals/joss-reviews#1774) and have a couple of questions

First, to echo the two points already mentioned by @harpolea:
-has the performance question been addressed? Since the equation is given explicitly it would be nice to have this substantiated quantitatively with a repex if possible
-again to echo the state of the field issue: you mention other tools using server side calculations - can you mention what some of these tools are and how they compare (performance wise)?
And in addition:
-I didn't see references in the paper - can you point these out if I'm just missing them somehow (perhaps they are somewhere else)
-regarding automated test cases: although the example notebooks are fairly in depth, they are quite specific to niche use-cases and maybe a more general 'set-up and testing' type tutorial could be added in addition to these, to show base functionality with sparse datasets as mentioned in the paper?
