### 2019-12-02 #1774 @KayleighRutherford
Hi @lpantano, I've completed this and wrote to the authors to address the remaining issues, but am now unable to edit my checklist? Can you advise on how to fix this (I am logged in and obviously accepted the invite as outlined but have since lost the ability to edit for some reason). Please let me know and I'll close this out. Thanks for waiting on me! Sorry this took some time
