### 2019-12-04 #1774 @KayleighRutherford
my comments are in issue #45 in the repo (sorry just realised I was supposed to create a new one..)
