### 2019-12-04 1e1ee0b
add section on timing results
