### 2019-12-04 b12e4b8
fix equation formatting, add backticks for packages
