+++ paper/paper.md
@@ -20,21 +20,21 @@ bibliography: paper.bib
 
 # Summary
 
-widgyts is a custom jupyter widget library to assist in interactive data
-visualization and exploration with yt. yt [@turk_yt:_2011] is a python 
+`widgyts` is a custom jupyter widget library to assist in interactive data
+visualization and exploration with `yt`. `yt` [@turk_yt_2011] is a python 
 package designed to read, process, and visualize multidimensional
-scientific data. yt allows users to ingest and visualize data from 
+scientific data. `yt` allows users to ingest and visualize data from 
 a variety of scientific domains with a nearly identical set of commands. Often,
 these datasets are large, sparse, complex, and located remotely. Creating a
 publication-quality figure of an area of interest for this data may take
 numerous exploratory visualizations and subsequent parameter-tuning events.
-The widgyts package allows for interactive exploratory visualization with yt, 
+The `widgyts` package allows for interactive exploratory visualization with `yt`, 
 enabling users to more readily determine which parameters and selections they
 need to best display their data. 
 
-The widgyts package is built on the ipywidgets [@grout_ipywidgets_2019] framework, which
-allows yt users to browse their data using a Jupyter notebook or a Jupyterlab
-instance. widgyts is developed on GitHub in the Data Exploration Lab organization. Issues,
+The `widgyts` package is built on the `ipywidgets` [@grout_ipywidgets_2019] framework, which
+allows `yt` users to browse their data using a Jupyter notebook or a Jupyterlab
+instance. `widgyts` is developed on GitHub in the Data Exploration Lab organization. Issues,
 questions, new feature requests, and any other relevant discussion can be found
 at the source code repository [@munk_widgyts_2019].
 
@@ -48,38 +48,41 @@ before the features are highlighted to effectively convey a message. If a
 dataset is located on a remote server, for each interaction a request is sent
 to the server, the server then calculates a new image, serializes it, and the
 image is sent back to the client. The total time to generate one image can
-generally be expressed as $T_{server}$, where 
+generally be expressed as $T_{\text{server}}$, where 
 
-$$t_{server} = t_{request} + t_{image calc, server} + t_{pull,image} + t_{display}.
+$$t_{\text{server}} = t_{\text{request}} + t_{\text{image calc, server}} + \
+t_{\text{pull, image}} + t_{\text{display}}.
 $$
 
-The total compute time spent on image generation is $T_{server} = n*t_{server}$, where $n$
+The total compute time spent on image generation is $T_{\text{server}} = n*t_{\text{server}}$, where $n$
 is the number of interactions with the figure. 
 
-Widgyts modifies this process by shifting image calculation to occur
+`widgyts` modifies this process by shifting image calculation to occur
 client-side. Rather than image serialization and calculation happening at a
-remote server, a portion of the original data is uploaded into the WebAssembly backend of
-widgyts. The time to calculate image client-side can be expressed as:
+remote server, a portion of the original data is uploaded into the `WebAssembly` backend of
+`widgyts`. The time to calculate image client-side can be expressed as:
 
-$$t_{client} = t_{request} + t_{pull,data} + t_{image calc, client} + t_{display}.$$
+$$t_{\text{client}} = t_{\text{request}} + t_{\text{pull,data}} + t_{\text{image calc, client}} \
++ t_{\text{display}}.$$
 
 Subsequent interactions ($n$) with the image only affect the final two terms of
 the equation, so 
 
-$$T_{client} = t_{request} + t_{pull,data} + n*[t_{image calc, client} + t_{display}].$$
+$$T_{\text{client}} = t_{\text{request}} + t_{\text{pull,data}} + \
+n*[t_{\text{image calc, client}} + t_{\text{display}}].$$
 
 Thus, this becomes advantageous as 
 
-$$ T_{client} < T_{server}
[…]
