### 2019-12-04 ad6dcbf
add sentence clarification suggested by @matthewturk
