### 2019-12-04 #1774 @munkm
@lpantano @KayleighRutherford and @harpolea -- thank you all for the reviews (and helping shepherd the paper). I've opened some PRs hopefully addressing the reviewer comments in the repository (data-exp-lab/widgyts#46 and data-exp-lab/widgyts#47). After they get reviewed and merged, I think things should be good!
