### 2019-12-05 e2c8b4a
minor rephrase of a few sections, fix spelling errors
