+++ paper/paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 # Summary
 
-`widgyts` is a custom jupyter widget library to assist in interactive data
+`widgyts` is a custom Jupyter widget library to assist in interactive data
 visualization and exploration with `yt`. `yt` [@turk_yt_2011] is a python 
 package designed to read, process, and visualize multidimensional
 scientific data. `yt` allows users to ingest and visualize data from 
@@ -58,7 +58,8 @@ The total compute time spent on image generation is $T_{\text{server}} = n*t_{\t
 is the number of interactions with the figure. 
 
 `widgyts` modifies this process by shifting image calculation to occur
-client-side. Rather than image serialization and calculation happening at a
+client-side in the browser. 
+Rather than image serialization and calculation happening on a
 remote server, a portion of the original data is uploaded into the `WebAssembly` backend of
 `widgyts`. The time to calculate image client-side can be expressed as:
 
@@ -75,7 +76,7 @@ Thus, this becomes advantageous as
 
 $$ T_{\text{client}} < T_{\text{server}}
 n*t_{\text{image calc, client}} + t_{\text{pull, data}} < n*[t_{\text{image_calc, server}} + \ 
-t_{\text{pull, image}}]
+t_{\text{pull, image}}] .
 $$
 
 The time to pull an image or data is dependent on the data size and the
@@ -94,13 +95,14 @@ resolution, such as one calculated on an adaptive mesh,
 with a fixed higher resolution will require recalculating and sending
 pixel values for a region that may only be represented by a single value. 
 
-# The WebAssembly backend
+# The WebAssembly Backend
 
 To allow for efficient data loading into the browser we chose to use Rust
 compiled to WebAssembly. The WebAssembly backing of `widgyts` allows for binary, zero-copy
 storage of loaded data on the client side, and WebAssembly has been designed to
 interface well with JavaScript. Further, the primitive structure of WebAssembly
-reduces the time to calculate the image in the browser. Finally, WebAssembly
+reduces the time to calculate the image in the browser, thus reducing the time
+to calculate the image client-side. Finally, WebAssembly
 is executed in a sandboxed environment separate from other processes 
 and is memory safe. 
 
@@ -108,34 +110,42 @@ While `yt` can access data at an arbitrary location within the dataset, `widgyts
 is structured to access any data within a 2D slice. Thus, only a slice of the
 data is uploaded client-side, not the entire dataset. For the large, sparse
 datasets that `widgyts` has been designed for, it would be infeasible to upload
-the entire dataset into the browser. However, a new slice in
-the third dimension will require a new data upload from the server so not all
+the entire dataset into the browser. A new slice in
+the third dimension will require an additional data upload from the server. Therefore,
+not all
 exploration of the dataset can be performed exclusively client-side.
 
 # Results
 
 The following image is a simple timing comparison between using `yt` with 
 the Jupyter widgets package `ipywidgets` 
-and using the `widgyts` package on the same dataset. 
+and using the `widgyts` package on the same dataset. The dataset is the
+IsolatedGalaxy dataset; a commonly used example in the yt documentation
+consisting of a galaxy simulation perfomed on an adaptive mesh. The dataset has
+variable resolution and is sparse near the domain boundaries. 
 A notebook is
 included in the `widgyts` repository [@munk_widgyts_2019] for one interested in replicating this
 analysis.
 
 -![A timing comparison between using `ipywidgets` with `yt` and `widgyts` on
-the IsolatedGalaxy dataset distributed with `yt`. Each timing point is based 
-on a number of panning interactions in x, averaged over 10 measurements. Each
-data point includes a 95% confidence interval.](timing.png)
[…]
