### 2019-12-06 0044d0d
fix equation issue
