### 2019-12-06 2ffc432
try to fix the figure caption issue
