+++ paper/paper.md
@@ -155,7 +155,7 @@ A notebook is
 included in the `widgyts` repository [@munk_widgyts_2019] for one interested in replicating this
 analysis.
 
--![A timing comparison between using `ipywidgets` with `yt` and `widgyts` on
+![A timing comparison between using `ipywidgets` with `yt` and `widgyts` on
 the IsolatedGalaxy dataset distributed with `yt`. The image generated by each 
 tool is 512x512 pixels. Each timing point is based 
 on a number of panning interactions in x, averaged over 10 measurements. The
