### 2019-12-06 7253b64
update matplotlib reference to include doi
