+++ paper/paper.bib
@@ -1,13 +1,17 @@
 
 @article{hunter_matplotlib_2007,
-	title = {Matplotlib: {A} 2D graphics environment},
-	volume = {9},
-	shorttitle = {Matplotlib},
-	number = {3},
-	journal = {Computing in science \& engineering},
-	author = {Hunter, John D.},
-	year = {2007},
-	pages = {90},
+  Author    = {Hunter, J. D.},
+  Title     = {Matplotlib: A 2D graphics environment},
+  Journal   = {Computing in Science \& Engineering},
+  Volume    = {9},
+  Number    = {3},
+  Pages     = {90--95},
+  abstract  = {Matplotlib is a 2D graphics package used for Python for
+  application development, interactive scripting, and publication-quality
+  image generation across user interfaces and operating systems.},
+  publisher = {IEEE COMPUTER SOC},
+  doi       = {10.1109/MCSE.2007.55},
+  year      = 2007
 }
 
 @article{turk_yt_2011,
