### 2019-12-06 #1774 @munkm
Hi @lpantano I just pushed a change that should fix the doi in that reference. Should I regenerate the PDF?
