### 2019-12-10 #1774 @lpantano
@KayleighRutherford can you confirm that all your questions are addressed and checked the items if yes, or ask for more if no? Thanks
