### 2019-12-17 1b4d13d
Ran javascript through js-beautify
