+++ js/lib/colormaps.js
@@ -3,26 +3,24 @@ var ipydatawidgets = require('jupyter-dataserializers');
 var _yt_tools = import('@data-exp-lab/yt-tools');
 var _ = require('lodash');
 var EXTENSION_VERSION = require('../package.json').version
-
 var CMapModel = widgets.WidgetModel.extend({
-
     defaults: function() {
-        return _.extend(widgets.WidgetModel.prototype.defaults.call(this), {
+        return _.extend(widgets.WidgetModel.prototype.defaults.call(
+            this), {
             _model_name: 'CMapModel',
             _model_module: '@data-exp-lab/yt-widgets',
             _model_module_version: EXTENSION_VERSION,
-
             cmaps: undefined,
             map_name: null,
             is_log: undefined,
-            min_val: undefined, 
+            min_val: undefined,
             max_val: undefined,
             generation: undefined
         });
     },
-
     initialize: function() {
-        widgets.WidgetModel.prototype.initialize.apply(this, arguments);
+        widgets.WidgetModel.prototype.initialize.apply(this,
+            arguments);
         this.map_name = this.get('map_name');
         this.is_log = this.get('is_log');
         this.min_val = this.get('min_val');
@@ -30,43 +28,40 @@ var CMapModel = widgets.WidgetModel.extend({
         this.generation = this.get('generation');
         this.data_array = null;
         this.image_array = null;
-
         this.setupListeners();
     },
-
-    normalize: function() {return _yt_tools.then(function(yt_tools) {
-        // normalizes a given buffer with a colormap name. Requires colormaps
-        // to be loaded in to wasm, so requires add_mpl_colormaps to be called 
-        // at this time.
-        //
-        var colormaps = this.get_cmaps(yt_tools);
-        colormaps.normalize(this.map_name, this.data_array, 
-          this.image_array, this.min_val, this.max_val, this.is_log);
-        // I sort of feel like this next line shouldn't be required if we 
-        // update the Python side, but whatever. 
-        // Updates the js side of image_array to our result. 
-        //this.image_array = array;
-
-        // this sync isn't working yet, so on the python side we can't 
-        // access it. 
-        // However, in order for the FRB to pick up that something changed 
-        // in the image array, this.set must be used.  
-        this.generation = this.generation + 1;
-        this.set('generation', this.generation);
-        this.save_changes();
-        return this.image_array;
-    }.bind(this)) },
-
+    normalize: function() {
+        return _yt_tools.then(function(yt_tools) {
+            // normalizes a given buffer with a colormap name. Requires colormaps
+            // to be loaded in to wasm, so requires add_mpl_colormaps to be called 
+            // at this time.
+            //
+            var colormaps = this.get_cmaps(yt_tools);
+            colormaps.normalize(this.map_name, this
+                .data_array, this.image_array, this
+                .min_val, this.max_val, this.is_log);
+            // I sort of feel like this next line shouldn't be required if we 
+            // update the Python side, but whatever. 
+            // Updates the js side of image_array to our result. 
+            //this.image_array = array;
+            // this sync isn't working yet, so on the python side we can't 
+            // access it. 
+            // However, in order for the FRB to pick up that something changed 
+            // in the image array, this.set must be used.  
+            this.generation = this.generation + 1;
+            this.set('generation', this.generation);
+            this.save_changes();
+            return this.image_array;
+        }.bind(this))
+    },
     get_cmaps: function(yt_tools) {
         // initializes the wasm colormaps module from yt tools and adds the 
         // arrays stored in the self.cmaps dict on the python side into
         // the colormaps object in wasm.
-        
[…]
