### 2019-12-24 #1774 @munkm
Ah! Thanks for the tip @lpantano. I manually updated the name and author list on Zenodo. I think it should match now.
