### 2020-01-02 #1774 @lpantano
Hi @openjournals/joss-eics, I think you can check this one now. happy new year!
