### 2020-01-07 #1774 @kthyng
Hi @munkm! I am going to finish up the processing of your paper.

I am going through your paper now. Can you look into the reference for Ipywidgets? Should the final author be Min RK or should it be his full name?
