### 2020-01-07 #1774 @kthyng
I see there is an issue with the references in the 2nd paragraph under "motivation" but I'm having trouble figuring out the source. For one thing, is there no author available for your "jupyter-matplotlib_2019" reference, nor "itkwidgets_2019", "bokeh_2019", and "bqplot_2019"?
