### 2020-01-07 #1774 @munkm
Hi @kthyng -- thank you for catching these! Min's name should indeed be longer. I'll update it and push again.

For the references in the 2nd paragraph, I used Zotero to import citation data on github repositories. None of these have a citation file, but there is a "publisher" information (e.g. Matplotlib developers) that was imported that could be used as authorship. When I use citeas it looks like this is what's used in place of individual authorship (e.g. https://citeas.org/cite/https://github.com/bokeh/bokeh).

Which would be better in this case? Duplicating the publisher information to the authorship information? Or copying first few contributor names in the repo to the author section of the citation? I'll update with whichever you think is best.
