+++ paper/paper.bib
@@ -66,7 +66,7 @@ archivePrefix = "arXiv",
 	url = {https://github.com/jupyter-widgets/ipywidgets},
 	urldate = {2019-09-06},
 	publisher = {Jupyter Widgets},
-	author = {Grout, Jason and Frederic, Jonathan and Corlay, Sylvain and RK,, Min},
+	author = {Grout, Jason and Frederic, Jonathan and Corlay, Sylvain and Ragan-Kelley, Min},
 	month = sep,
 	year = {2019},
 	note = {original-date: 2015-04-17T20:29:25Z},
