### 2020-01-07 #1774 @kthyng
@munkm Whoa I didn't "citeas" existed! Very cool! I think if a project hasn't given a preferred citation and doesn't have a clear paper to cite, using the "citeas" approach of the publisher as author will probably look best. Can you try that out and see how it looks inline? You can test here if you want: https://whedon.theoj.org/
