### 2020-01-24 307f24e
update references for itk, itkwidgets, bokeh, and bqplot
