+++ paper/paper.bib
@@ -74,52 +74,100 @@ archivePrefix = "arXiv",
 }
 
 @misc{jupyter-matplotlib_2019,
-	title = {matplotlib/jupyter-matplotlib},
+	title = {jupyter-matplotlib},
 	copyright = {BSD-3-Clause},
 	url = {https://github.com/matplotlib/jupyter-matplotlib},
 	abstract = {Matplotlib Jupyter Extension. Contribute to matplotlib/jupyter-matplotlib development by creating an account on GitHub.},
 	urldate = {2019-11-19},
-	publisher = {Matplotlib Developers},
+	publisher = {{Matplotlib Developers}},
 	month = nov,
 	year = {2019},
 	note = {original-date: 2016-07-21T14:38:56Z},
 	keywords = {jupyterlab-extension}
 }
 
-@misc{itkwidgets_2019,
-	title = {{InsightSoftwareConsortium}/itkwidgets},
-	copyright = {Apache-2.0},
-	url = {https://github.com/InsightSoftwareConsortium/itkwidgets},
-	abstract = {Interactive Jupyter widgets to visualize images, point sets, and meshes in 2D and 3D},
-	urldate = {2019-11-19},
-	publisher = {Insight Software Consortium},
-	month = nov,
-	year = {2019},
-	note = {original-date: 2018-02-15T01:43:39Z},
-	keywords = {image-analysis, imaging, insight-toolkit, itk, jupyter, jupyter-widget, jupyterlab-extension, kitware, medical-visualization, meshes, open-science, open-source, point-cloud, python, reproducible-research, scientific-computing, scientific-visualization}
-}
-
 @misc{bokeh_2019,
-	title = {bokeh/bokeh},
+	title = {Bokeh: Python library for interactive visualization},
 	copyright = {BSD-3-Clause},
-	url = {https://github.com/bokeh/bokeh},
+	url = {https://bokeh.org/},
 	abstract = {Interactive Data Visualization in the browser, from  Python},
 	urldate = {2019-11-19},
+	author = {{Bokeh Development Team}},
 	publisher = {Bokeh},
-	month = nov,
-	year = {2019},
-	note = {original-date: 2012-03-26T15:40:01Z},
-	keywords = {bokeh, data-visualisation, interactive-plots, javascript, plots, plotting, python, visualization}
 }
 
 @misc{bqplot_2019,
-	title = {bloomberg/bqplot},
+	title = {bqplot: Plotting for Jupyter},
 	copyright = {Apache-2.0},
 	url = {https://github.com/bloomberg/bqplot},
-	abstract = {Plotting library for IPython/Jupyter Notebooks. Contribute to bloomberg/bqplot development by creating an account on GitHub.},
+	abstract = {Plotting library for IPython/Jupyter Notebooks.},
 	urldate = {2019-11-19},
+	author = {Corlay, Sylvain and Sunkara, Srinivas and Madeka, Dhruv and Menegaux, Romain and Cherukuri, Chakri and Grout, Jason and Mabille, Johan},
 	publisher = {Bloomberg},
 	month = nov,
 	year = {2019},
 	note = {original-date: 2015-09-26T04:02:18Z}
+}
+
+@software{itk_2019,
+  author       = {Luis Ibanez and
+                  Bill Lorensen and
+                  Matt McCormick and
+                  Brad King and
+                  Daniel Blezek and
+                  Hans Johnson and
+                  Bradley Lowekamp and
+                  jjomier and
+                  Kitware, Inc. and
+                  Gaëtan Lehmann and
+                  Arnaud Gelas and
+                  karthikkrishnan and
+                  Mathieu Malaterre and
+                  Bill Hoffman and
+                  Dženan Zukić and
+                  Will and
[…]
