### 2020-01-24 4b199b0
update authors of jupyter-matplotlib
