+++ paper/paper.bib
@@ -80,6 +80,19 @@ archivePrefix = "arXiv",
 	abstract = {Matplotlib Jupyter Extension. Contribute to matplotlib/jupyter-matplotlib development by creating an account on GitHub.},
 	urldate = {2019-11-19},
 	publisher = {{Matplotlib Developers}},
+    author = {Sylvain Corlay and 
+              Steven Silvester and
+              martinRenou and
+              Kyle Boone and
+              Thomas A Caswell and
+              Jens Hedegaard Nielsen and
+              Erik Sundell and
+              Greg and
+              Jason Grout and
+              Matt McCormick and
+              Nathan Dunfield and
+              Tim Paine and
+              Todd},
 	month = nov,
 	year = {2019},
 	note = {original-date: 2016-07-21T14:38:56Z},
