### 2020-01-24 d8552a8
update bokeh year in bibfile, fix itk ref in paper
