+++ paper/paper.bib
@@ -99,13 +99,13 @@ archivePrefix = "arXiv",
 	keywords = {jupyterlab-extension}
 }
 
-@misc{bokeh_2019,
+@Manual{bokeh_2019,
 	title = {Bokeh: Python library for interactive visualization},
 	copyright = {BSD-3-Clause},
 	url = {https://bokeh.org/},
-	abstract = {Interactive Data Visualization in the browser, from  Python},
 	urldate = {2019-11-19},
-	author = {{Bokeh Development Team}},
+	year = {2019},
+    author = {{Bokeh Development Team}},
 	publisher = {Bokeh},
 }
 
