### 2020-01-24 #1774 @munkm
Hi @kthyng! I just heard about `citeas` recently, too! I'm a big fan! ❤️

Thank you for linking the paper preview service! It's so helpful.

I've fixed the references based on opening issues in a few repositories to check what the package developers preferred. I think I've fixed most of the weird citation issues now and they look good in the preview service to me. Can you let me know if everything is ok on your end?
