### 2020-01-29 #1774 @kthyng
Hi @munkm! I just read through your paper, and have just a couple of typos in this [PR](https://github.com/data-exp-lab/widgyts/pull/51)
