### 2020-01-29 #1774 @munkm
Thanks for catching them 🙏, it's really appreciated! Your fixes have been merged.
