### 2020-01-29 #1774 @kthyng
Hi all! This submission is almost wrapped up from my end, but I just perused the checklists and other items above and I see that @KayleighRutherford gave approval (though wasn't able to check the final items off due to permissions) but I'm not sure we heard a final word from @harpolea, particularly given an open issue [#45](https://github.com/data-exp-lab/widgyts/issues/45). I'd like to make sure this issue is satisfactorily completed before finalizing this submission.
