### 2020-01-29 data-exp-lab/widgyts#45 @kthyng
@harpolea can you give an update here as to the status of your suggestions? Thanks.
