### 2020-01-29 data-exp-lab/widgyts#45 @harpolea
The additions to the paper look great to me, so I'm happy to close this issue.
