### 2020-01-29 #1774 @harpolea
I'm happy to sign off on this - the modifications to the paper have addressed the suggestions I made, so all looks good to me!
